start working on edit module

+++ src/app/queries/crateModule.ts
@@ -1,7 +1,7 @@
 import gql from 'graphql-tag';
 
 export default gql`
-    mutation AddModule($name: String!, $bodytext: String!, $price: Int!, $groupId: String) {
+    mutation AddModule($name: String!, $bodytext: String!, $price: Float!, $groupId: String) {
         addModule(name: $name, bodytext: $bodytext, price: $price, groupId: $groupId ){
             name
         }
